requirements.txt
AwesomeTkinter==2021.11.8
Pillow==9.5.0
pyperclip==1.8.2
